+++ frontend/src/components/datepicker/NowDatePicker.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
+const selectedDate = defineModel<Date>({required: true});
 defineProps<{ label: string }>();
-const emit = defineEmits(['change']);
 </script>
 
 <template>
-    <div class="mt-3">
-        <button class="btn btn-primary w-100" @click="emit('change', new Date())">
+    <div class="mt-5 mb-5">
+        <button class="btn btn-primary w-100" @click="selectedDate = new Date()">
             Set {{ label.toLowerCase() }} to now
         </button>
     </div>
